+++ src/components/explore/utils.tsx
@@ -1,3 +1,6 @@
+import { DetailsSchema } from 'components/context'
+import { TonsWithAddl } from './types'
+
 // Create nice listy thing w/truncation like
 // a, b, c, d, e...
 // a, b, c, d
@@ -20,4 +23,91 @@ export const pluralTextIfNeeded = (length: number, text = 'item'): string => {
   if (length === 1) return `${length} ${text}`
 
   return `${length} ${text}s`
+}
+
+// Mid-level formulas are pretty consistent, except:
+//  1. /Explore/Landing does not need a formula
+//  2. /Explore/{anything with array field in Language}/:value must check for
+//     value existence and then a "contains" since there does not appear to be a
+//     way to filter arrays in Airtable.
+export const prepFormula = (
+  field: keyof DetailsSchema,
+  value?: string
+): string => {
+  if (field === 'Language') return '' // /Explore/Language
+  if (!value) return "{languages} != ''" // e.g. /Explore/Country
+
+  const midLevelArrayFields = [
+    'Country',
+    'Macrocommunity',
+    'Neighborhood',
+    'Town',
+  ]
+
+  // Neighborhood instance, e.g. /Explore/Neighborhood/Astoria should include
+  // additional neighborhoods in the query.
+  if (value && field === 'Neighborhood')
+    return `AND(
+      {${field}} != '',
+      FIND(
+        '${value}',
+        CONCATENATE(
+          ARRAYJOIN({${field}}),
+          ARRAYJOIN({addlNeighborhoods})
+        )
+      ) != 0
+    )`
+
+  if (value && midLevelArrayFields.includes(field))
+    return `AND({${field}} != '', FIND('${value}', ARRAYJOIN({${field}})) != 0)`
+
+  if (value) return `{${field}} = '${value}'`
+
+  return ''
+}
+
+// Mid-level fields are consistent except a few tables need an extra field
+export const prepFields = (tableName: keyof DetailsSchema): string[] => {
+  if (tableName === 'Language')
+    return [
+      'Endonym',
+      'name',
+      'Primary Locations',
+      'worldRegionColor',
+      'addlNeighborhoods',
+    ]
+
+  const landingFields = ['name', 'languages']
+  const addlFields: {
+    [key: string]: string[]
+  } = {
+    'World Region': [...landingFields, 'icon-color'],
+    Country: [...landingFields, 'src_image'],
+    Neighborhood: [...landingFields, 'Additional Languages'],
+  }
+
+  if (addlFields[tableName] !== undefined) return addlFields[tableName]
+
+  return landingFields
+}
+
+export const getUniqueInstances = (
+  field: string,
+  row: TonsWithAddl,
+  value?: string
+): string[] => {
+  let uniqueInstances = []
+
+  if (field === 'Neighborhood' && !value) {
+    uniqueInstances = [
+      ...(row.languages || []),
+      ...(row['Additional Languages'] || []),
+    ]
+  } else if (row.languages) {
+    uniqueInstances = row.languages
+  } else {
+    uniqueInstances = row['Primary Locations']
+  }
+
+  return uniqueInstances
 }
